Add options dialog for setting the colors
This brings in a dependency on SciJava Common for its options service,
as well as a dependency on the SciJava AWT UI for its AWTColors
utility class.

Initial version only lets you set the first eight colors.

And settings are not hooked up to the actual Cell Counter plugin yet.
But will be shortly.